requirements.txt
streamlit>=1.24.0
supabase>=1.0.3
python-dotenv>=1.0.0
openai>=1.3.0
requests>=2.31.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
pillow>=10.0.0
